+++ src/fibery/builders.py
@@ -4,6 +4,7 @@ from uuid import uuid4
 
 from .entity_model import FiberyBaseModel
 from .fibery_models import FiberyCommand
+from .utils import CollectionOperation
 
 
 class QueryBuilder:
@@ -107,6 +108,7 @@ class EntityBuilder:
     def prepare_command(type_name: str, data: FiberyBaseModel) -> tuple[str, FiberyCommand]:
         entity_id = str(uuid4())
         command = FiberyCommand(
+            command='fibery.entity/create',
             args={
                 'type': type_name,
                 'entity': {
@@ -116,3 +118,35 @@ class EntityBuilder:
             }
         )
         return entity_id, command
+
+
+    @staticmethod
+    def prepare_update_command(type_name: str, entity_id: str, updates: dict[str, Any]) -> FiberyCommand:
+        return FiberyCommand(
+            command='fibery.entity/update',
+            args={
+                'type': type_name,
+                'entity': {
+                    'fibery/id': entity_id,
+                    **updates
+                }
+            }
+        )
+
+    @staticmethod
+    def prepare_collection_command(
+            type_name: str,
+            entity_id: str,
+            field: str,
+            item_ids: list[str],
+            operation: CollectionOperation,
+    ) -> FiberyCommand:
+        return FiberyCommand(
+            command=operation.command,
+            args={
+                'type': type_name,
+                'entity': {'fibery/id': entity_id},
+                'field': field,
+                'items': [{'fibery/id': item_id} for item_id in item_ids]
+            }
+        )