+++ CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.3] - 2024-02-23
+
+### Changed
+- Changed upload_file method - remove curl workaround
+
+## [0.1.2] - 2024-01-29
+
+### Added
+- New methods to handle collections and files
+
+## [0.1.1] - 2024-01-27
+
+### Added
+- New update_entity method
+- New find_and_update_entity method
+- Refactor EntityBuilder
+- Add default fibery_id field to FiberyBaseModel
+- Add more logs for fibery service
+
 ## [0.1.0] - 2024-01-27
 
 ### Added